update half -> float

+++ PowerPBS/Lib/ShadowCasterPass.hlsl
@@ -7,17 +7,17 @@
 
 
 struct v2f{
-    half2 uv:TEXCOORD0;
-    half4 pos:SV_POSITION;
+    float2 uv:TEXCOORD0;
+    float4 pos:SV_POSITION;
 };
 
-half3 _LightDirection;
+float3 _LightDirection;
 
 //--------- shadow helpers
-half4 GetShadowPositionHClip(appdata_full input){
-    half3 worldPos = mul(unity_ObjectToWorld,input.vertex).xyz;
-    half3 worldNormal = UnityObjectToWorldNormal(input.normal);
-    half4 positionCS = UnityWorldToClipPos(ApplyShadowBias(worldPos,worldNormal,_LightDirection));
+float4 GetShadowPositionHClip(appdata_full input){
+    float3 worldPos = mul(unity_ObjectToWorld,input.vertex).xyz;
+    float3 worldNormal = UnityObjectToWorldNormal(input.normal);
+    float4 positionCS = UnityWorldToClipPos(ApplyShadowBias(worldPos,worldNormal,_LightDirection));
     #if UNITY_REVERSED_Z
         positionCS.z = min(positionCS.z, positionCS.w * UNITY_NEAR_CLIP_VALUE);
     #else
@@ -39,10 +39,10 @@ v2f vert(appdata_full input){
     return output;
 }
 
-half4 frag(v2f input):SV_Target{
+float4 frag(v2f input):SV_Target{
     #if defined(_ALPHA_TEST)
     if(_AlphaTestOn){
-        half4 tex = SAMPLE_TEXTURE2D(_MainTex,sampler_MainTex,input.uv);
+        float4 tex = SAMPLE_TEXTURE2D(_MainTex,sampler_MainTex,input.uv);
         clip(tex.a - _Cutoff);
     }
     #endif
